refactor(work-state): extract graph-validation, task-readiness, parallel-workers into work-state/ submodules
